pyproject.toml
[project]
name = "runner2d"
version = "0.1.0"
description = "Top-down survival prototype scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.7",
]

[project.scripts]
runner2d = "game.run:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
